requirements.txt
stable_baselines3[extra]
pettingzoo
mujoco>=3.1.0
pyyaml
scipy
scikit-image
imageio
opencv-python
fcmaes

hydra-core @ git+https://github.com/AaronYoung5/hydra.git
hydra_zen @ git+https://github.com/mit-ll-responsible-ai/hydra-zen.git
omegaconf @ git+https://github.com/omry/omegaconf.git

hydra-nevergrad-sweeper @ git+https://github.com/AaronYoung5/hydra.git@main#subdirectory=plugins/hydra_nevergrad_sweeper
hydra-optuna-sweeper @ git+https://github.com/AaronYoung5/hydra.git@main#subdirectory=plugins/hydra_optuna_sweeper
hydra-ax-sweeper @ git+https://github.com/AaronYoung5/hydra.git@main#subdirectory=plugins/hydra_ax_sweeper
hydra-submitit-launcher @ git+https://github.com/AaronYoung5/hydra.git@main#subdirectory=plugins/hydra_submitit_launcher
